Changed the naming of all files and folders to use a period, '.', when indicating that the next character is an uppercase character. This period, '.', in files also indicates a - or a _ in the uri. I also changed it so that branch controllers dont exist anymore and are renamed to application controllers. Drivers have been renamed to db.mysql.driver.class changing the class name to DB_Mysql_Driver.
All classes now have to be named after their type and their branch following this scheme: [Name]_[Branch]_[Type] so if i was naming a helper in the main structure as test.php the class name would be Test_Helper and if the helper was in the admin branch the class name would be Test_Admin_Helper.

Deleted the system class and made it so that all classes, e.g. helpers, models, plugins, controllers, are loaded in using the name of the class using the new keyword. So if I was to load the test helper as indicated above I would use this syntax: $helper = new Test_Helper(); allowing me to also use constructors. Error handling for classes has been enhanced as a result of this change so that any class can be validated for its existence even if it is being extended and not called directly by the programmer.

The designer class has been deleted and its main functions have been ported over to the controller class.

The uri map now supports the ability to prepend uri elements before the controller uri element. In order to get the results you would desire you will need to have a validator for it which is done by defining a default with an array and having the second element in the array be a function that was created using the create_function method as shown here: "lang" => array("en", create_function('$element', 'return preg_match("/(en|sp)/i", $element);')). This will validate if the uri value at the position of "lang" is what is being looked for or not. If not then the default if used as defined by the first position in the array and if it is what is being looked for then the uri value is assigned as lang and the rest of the uri is assigned appropriately.

The uri has been made a bit more ridged so that it now requires that the uri map has a controller and a view key defined in it. This has made a lot of code much simpler in the controller class, the config class, and in the evergreen class.

The error handler now sends headers for more codes such as, 301, 404 and 500. The sending of the headers now also happens when the error is being processed and not only if a url is defined. Also the error class now will be able to route you to a url that is not in the framework, so if you want to do a 301 and set a location to a page not on your site you would be able to do so. The error handler also does a check for the existence of its default files and if they dont exist it will just echo out the value of its default errors. The error handler now also has sprintf abilities in its messages using the messageArgs array key in params. The sprintf features are enhanced to have the ability to use array keys as well in this format: %(key)s it also has the ability to use any of the defined configuration variables such as system version: %(System.version)s so that error messages can be more robust.

The formhandler has been changed so that it doesn’t use preg_replace with the e modifier anymore and rather used preg_replace_callback.

The factory class has been trimmed down in preparation for its deletion when the new models are completed.

Various other fixes have been made also along with some cleanup of variables and general clean up of code to make processing quicker and easier.